requirements.txt
fastapi[standard]
aiokafka
python_dotenv
